Share deterministic lexical normalization
