+++ crates/compass-model/Cargo.toml
@@ -18,6 +18,7 @@ rmp-serde.workspace = true
 rayon.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
+unicode-normalization.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
