+++ crates/compass-model/src/lib.rs
@@ -8,6 +8,7 @@ mod document;
 mod error;
 mod graph;
 pub mod identity;
+mod lexical;
 pub mod provenance;
 pub mod query_contract;
 mod query_index;
@@ -16,6 +17,7 @@ mod validation;
 pub use document::{EdgeRecord, GraphDocument, NodeRecord};
 pub use error::GraphError;
 pub use graph::{EdgeIndex, Graph, NodeIndex};
+pub use lexical::{canonical_code_token, identifier_tokens, strip_diacritics};
 pub use query_index::{QueryIndex, SchemaFingerprint, cypher_node_label, cypher_relationship_type};
 pub use validation::{
     CodeGraphValidationError, CodeGraphValidationReport, ExtractionValidationError,
